+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+## [1.27.0]
+
+- Add plot for TSC timestamp delta
+- Add timestamp statistics to TSC metadata
+- Add show/hide all button for `[bool]` data series
+
 ## [1.26.0]
 
 - Always try to fetch WASM bundle, fallback to cached if fetch fails